backend/server/requirements.txt
Django==5.2.11
djangorestframework>=3.15.2
django-allauth==0.63.3
django-money==3.5.4
django-invitations==2.1.0
drf-yasg==1.21.4
django-cors-headers==4.4.0
coreapi==2.3.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pillow==12.1.1
whitenoise==6.9.0
django-resized==1.0.3
django-geojson==4.2.0
setuptools==79.0.1
gunicorn==23.0.0
qrcode==8.0
slippers==0.6.2
django-allauth-ui==1.7.0
django-widget-tweaks==1.5.0
django-ical==1.9.2
icalendar==6.1.0
ijson==3.3.0
tqdm==4.67.1
overpy==0.7
publicsuffix2==2.20191221
geopy==2.4.1
psutil==6.1.1
geojson==3.2.0
gpxpy==1.6.2
pymemcache==4.0.0
legacy-cgi==2.6.3
requests>=2.31.0
aiosmtpd==1.4.6
